pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "visualizer"
version = "1.0.0"
description = "AI 모델 연구 & 제품 개발 통합 도구"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.8.0",
    "opencv-python>=4.10.0",
    "numpy>=1.26.0",
    "onnxruntime>=1.17.0",
    "psutil>=5.9.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
charts = ["matplotlib>=3.8.0", "scikit-learn>=1.4.0", "openpyxl>=3.1.0"]
umap = ["umap-learn>=0.5.0"]
all = ["matplotlib>=3.8.0", "scikit-learn>=1.4.0", "openpyxl>=3.1.0", "umap-learn>=0.5.0"]
dev = ["pytest>=8.0.0"]

[project.scripts]
visualizer = "run:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
